Include headers/lib from FreeTDS 0.91 for linux x64
FreeTDS Configuration
---------------------

export CC="gcc -fPIC"
./configure --disable-odbc --disable-apps --disable-server
    --disable-pool --with-tdsver=7.1 --prefix=/some/tmp/prefix

I had to force the -fPIC option to GCC in order to be able to link
libsybdb.a.  Before doing that, I was getting an error like the
one described here:

http://www.gentoo.org/proj/en/base/amd64/howtos/index.xml?part=1&chap=3

From that article:

    In this case, globally adding -fPIC to C[XX]FLAGS resolves the
    issue, although this practice is discouraged because the
    executables end up being PIC-enabled, too.

Since we are doing a special build and only need the library, not the
applications, it seems like a safe bet.