Initial support for connection to the Azure service.
Implement it by sending the name of the DB in the LOGINREC structure
passed to dbopen().

A 'azure' boolean argument is added to _mssql.connect() (and
_mssql.MSSQLConnection.__init__()) and to pymssql.connect().

Support by FreeTDS is already available since version 0.91

Credit goes to author of TinyTDS Ken Collins and collaborators
because this is heavily inspired by the Azure support there.